pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eda-rule-engine"
version = "0.1.0"
description = "Data Accuracy Test Rule Engine for automated validation of business logic rules"
authors = [{name = "EDA Team", email = "team@eda.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.1.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.17.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
eda-rule-engine = "eda_rule_engine.cli.main:app"

[project.urls]
Homepage = "https://github.com/your-org/eda-rule-engine"
Documentation = "https://github.com/your-org/eda-rule-engine/docs"
Repository = "https://github.com/your-org/eda-rule-engine"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true